+++ src/Messaging.Mongo.Client/MessagingClientBuilderExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using Assistant.Net.Messaging.Abstractions;
 using Assistant.Net.Messaging.Options;
 using Assistant.Net.Messaging.Serialization;
-using Microsoft.Extensions.Configuration;
 using System;
 
 namespace Assistant.Net.Messaging
@@ -12,38 +11,10 @@ namespace Assistant.Net.Messaging
     public static class MessagingClientBuilderExtensions
     {
         /// <summary>
-        ///     Configures the messaging client to connect a MongoDB database from a client.
-        /// </summary>
-        public static MessagingClientBuilder UseMongo(this MessagingClientBuilder builder, string connectionString) =>
-            builder.UseMongo(o => o.ConnectionString = connectionString);
-
-        /// <summary>
-        ///     Configures the messaging client to connect a MongoDB database from a client.
-        /// </summary>
-        public static MessagingClientBuilder UseMongo(this MessagingClientBuilder builder, Action<MongoOptions> configureOptions)
-        {
-            builder.Services
-                .AddMongoClientFactory()
-                .ConfigureMongoOptions(MongoOptions.ClientName, configureOptions);
-            return builder;
-        }
-
-        /// <summary>
-        ///     Configures the messaging client to connect a MongoDB database from a client.
-        /// </summary>
-        public static MessagingClientBuilder UseMongo(this MessagingClientBuilder builder, IConfigurationSection configuration)
-        {
-            builder.Services
-                .AddMongoClientFactory()
-                .ConfigureMongoOptions(MongoOptions.ClientName, configuration);
-            return builder;
-        }
-
-        /// <summary>
         ///     Registers remote MongoDB based handler of <typeparamref name="TMessage" /> from a client.
         /// </summary>
         /// <remarks>
-        ///     Pay attention, it requires calling one of <see cref="UseMongo(MessagingClientBuilder,string)"/>.
+        ///     Pay attention, it requires calling one of UseMongo method.
         /// </remarks>
         /// <typeparam name="TMessage">Specific message type to be handled remotely.</typeparam>
         /// <exception cref="ArgumentException"/>
@@ -54,7 +25,7 @@ namespace Assistant.Net.Messaging
         ///     Registers remote MongoDB based handler of <paramref name="messageType" /> from a client.
         /// </summary>
         /// <remarks>
-        ///     Pay attention, it requires calling one of <see cref="UseMongo(MessagingClientBuilder,string)"/>.
+        ///     Pay attention, it requires calling one of UseMongo method.
         /// </remarks>
         /// <exception cref="ArgumentException"/>
         public static MessagingClientBuilder AddMongo(this MessagingClientBuilder builder, Type messageType)