+++ tests/Messaging.Mongo.Tests/Fixtures/MessagingClientFixture.cs
@@ -34,7 +34,7 @@ namespace Assistant.Net.Messaging.Mongo.Tests.Fixtures
             {
                 o.Handlers.Clear();
                 foreach (var handlerInstance in handlerInstances)
-                    o.AddLocalHandler(handlerInstance);
+                    o.AddHandler(handlerInstance);
             });
             clientSource.Configurations.Add(o =>
             {